backend/pyproject.toml
[project]
name = "agentic-sql-backend"
version = "0.1.0"
description = "FastAPI backend for the agentic SQL educational demo"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.52.0",
    "pandas>=2.2",
    "openpyxl>=3.1",
    "pyarrow>=15.0",
    "duckdb>=0.10",
    "numpy>=1.26",
    "ollama>=0.3",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "pyright>=1.1.407",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

